aula 6 app lista de jogos, Professor Celso Furtado

+++ app/src/main/java/adeniltonmorais/android/recicleview/model/Jogo.kt
@@ -3,6 +3,6 @@ package adeniltonmorais.android.recicleview.model
 data class Jogo(
     var codigo:Int,
     var nomeJogo:String,
-    var notaJogo:Int,
+    var notaJogo:Float,
     var console:String
 )
